+++ src/main/java/dev/nullpanic/tamagotchitelegrambot/service/PetServiceImpl.java
@@ -5,9 +5,12 @@ import dev.nullpanic.tamagotchitelegrambot.persist.repository.PetRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 public class PetServiceImpl implements PetService {
 
+    private final int HungrinessIncrement = 20;
     private final PetRepository petRepository;
 
     @Autowired
@@ -16,7 +19,34 @@ public class PetServiceImpl implements PetService {
     }
 
     @Override
-    public Pet createPet(Pet pet) {
+    public Pet createPet(Long chatId, String petName) {
+        Pet pet = new Pet();
+        pet.setChatId(chatId);
+        pet.setActive(true);
+        pet.setName(petName);
+        pet.setDefaultStats();
+
         return petRepository.save(pet);
+    }
+
+    @Override
+    public Pet savePet(Pet pet) {
+        return petRepository.save(pet);
+    }
+
+    @Override
+    public void feedPet(Pet pet) {
+        int newHungriness = pet.getHungriness() + HungrinessIncrement;
+
+        if (newHungriness > 100) {
+            newHungriness = 100;
+        }
+
+        pet.setHungriness(newHungriness);
+    }
+
+    @Override
+    public Optional<Pet> findByNameAndChatId(String name, Long chatId) {
+        return petRepository.findByNameAndChatId(name, chatId);
     }
 }